backend/requirements.txt
asgiref==3.9.1
black==25.1.0
certifi==2024.8.30
click==8.2.1
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
ruff==0.12.12
sqlparse==0.5.3
